Added formatting/templating with handlebars
