--- index.html
@@ -1,136 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Benny's Clicker</title>
-
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Poppins&display=swap" rel="stylesheet">
-
-    <link rel="stylesheet" href="styles.css">
-    <script src="index.js" charset="utf-8" defer></script>
-</head>
-<body>
-    <header>
-        <h1 class="site-title"><a href="#">Benny's Clicker</a></h1>
-
-        <!-- <div class="right-aligned">
-            <h1 class="right-text"><a href="#">Save</a></h1>
-            <h1 class="right-text"><a href="#">Leaderboard</a></h1>
-        </div> -->
-    </header>
-
-    <main class="content">
-        <div class="clicker">
-            <p id="wood-number">Wood: 0</p>
-    
-            <div class="button-div">
-                <img id="wood-button" src="benny.png">
-            </div>
-            
-            <div class="clicker-button-container">
-                <img id = "axe" src = "axe1.png">
-                  <button id = "axe-button">Buy Axe</button>
-            </div>
-            <div class="auto-clicker-button-container">
-                <img id = "auto-axe" src = "axe1.png">
-                <button id = "auto-axe-button">Buy Auto Axe</button>
-            </div>
-            
-        </div>
-        <div class="store">
-            <h1 class="store-title">Store</h1>
-            <div class="post" data-price="500">
-                <div class="post-contents">
-                  <div class="post-image-container">
-                    <img src="https://i.pinimg.com/originals/07/e0/8c/07e08c1195f197dab954a89723339b3b.jpg" alt="Super nice laptop">
-                  </div>
-                  <div class="post-info-container">
-                    <a href="#" class="post-title">Super nice laptop</a> <span class="post-price">$500</span>
-                  </div>
-                </div>
-              </div>
-      
-              <div class="post" data-price="20">
-                <div class="post-contents">
-                  <div class="post-image-container">
-                    <img src="https://p1.pxfuel.com/preview/998/294/512/meadow-yellow-rain-coat-mountains.jpg" alt="Yellow Raincoat">
-                  </div>
-                  <div class="post-info-container">
-                    <a href="#" class="post-title">Yellow raincoat</a> <span class="post-price">$20</span>
-                  </div>
-                </div>
-              </div>
-      
-              <div class="post" data-price="100">
-                <div class="post-contents">
-                  <div class="post-image-container">
-                    <img src="http://static-21.sinclairstoryline.com/resources/media/9e359e52-2d79-4aa6-a43b-2ba534b9db0c-large16x9_378b19aa394049a4a9ca7101e4e75be4voodoodoughnutrecord660.jpg?1463527394583" alt="Very large pile of interesting donuts">
-                  </div>
-                  <div class="post-info-container">
-                    <a href="#" class="post-title">Very large pile of interesting donuts</a> <span class="post-price">$100</span>
-                  </div>
-                </div>
-              </div>
-      
-              <div class="post" data-price="250">
-                <div class="post-contents">
-                  <div class="post-image-container">
-                    <img src="http://static.legalsolutions.thomsonreuters.com/product_photos/p40307624-141728L.jpg" alt="Oregon laws">
-                  </div>
-                  <div class="post-info-container">
-                    <a href="#" class="post-title">Oregon laws</a> <span class="post-price">$250</span>
-                  </div>
-                </div>
-              </div>
-      
-              <div class="post" data-price="1">
-                <div class="post-contents">
-                  <div class="post-image-container">
-                    <img src="http://i.huffpost.com/gen/1573833/images/o-SUNSHINE-facebook.jpg" alt="Sunshine">
-                  </div>
-                  <div class="post-info-container">
-                    <a href="#" class="post-title">Sunshine</a> <span class="post-price">$1</span>
-                  </div>
-                </div>
-              </div>
-      
-              <div class="post" data-price="99">
-                <div class="post-contents">
-                  <div class="post-image-container">
-                    <img src="http://www.portlandhipster.com/wp-content/uploads/2015/09/hipsters.jpg" alt="Hipsters">
-                  </div>
-                  <div class="post-info-container">
-                    <a href="#" class="post-title">Hipsters</a> <span class="post-price">$99</span>
-                  </div>
-                </div>
-              </div>
-      
-              <div class="post" data-price="20000000">
-                <div class="post-contents">
-                  <div class="post-image-container">
-                    <img src="https://upload.wikimedia.org/wikipedia/commons/thumb/1/1a/102707-Oregon-AutzenStadium-ext.jpg/1200px-102707-Oregon-AutzenStadium-ext.jpg" alt="Football stadium (we kind of want to build a new one)">
-                  </div>
-                  <div class="post-info-container">
-                    <a href="#" class="post-title">Football stadium (we kind of want to build a new one)</a> <span class="post-price">$20000000</span>
-                  </div>
-                </div>
-              </div>
-      
-              <div class="post" data-price="10000">
-                <div class="post-contents">
-                  <div class="post-image-container">
-                    <img src="https://farm2.staticflickr.com/1258/1366422925_25663f44c8_z.jpg" alt="Very nice farmer's market">
-                  </div>
-                  <div class="post-info-container">
-                    <a href="#" class="post-title">Very nice farmer's market</a> <span class="post-price">$10000</span>
-                  </div>
-                </div>
-              </div>
-        </div>
-    </main>
-</body>
-</html>
