+++ public/styles.css
@@ -1,0 +1,157 @@
+/* Disable scrolling */
+html, body {margin: 0; height: 100%; overflow: hidden}
+
+body {
+    font-family: 'Poppins', sans-serif;
+    margin: 0;
+}
+
+header {
+    background-color: #000;
+    display: flex;
+    align-items: center;
+    position: relative;
+}
+
+header a {
+    color: #fff;
+}
+
+a {
+    color: black;
+    text-decoration: none;
+}
+
+.site-title, .right-text {
+    font-size: 24px;
+    font-weight: 100;
+    margin: 0 20px 0 20px;
+    padding-top: 20px;
+    padding-bottom: 10px;
+}
+
+/* .right-text {
+    text-align: right;
+} */
+  
+  .site-title a:hover {
+    text-decoration: none;
+}
+
+.content {
+    display: flex;
+}
+
+.clicker {
+    display: inline-block;
+    min-width: 50%;
+    max-width: 50%;
+    margin: 0 10px auto 10px;
+    padding: 20px;
+    border-radius: 3px;
+}
+
+#wood-number {
+    font-size:60px;
+    font-weight: bold;
+}
+
+.button-div {
+    /* Center image */
+    display: flex;
+    justify-content: center;
+    margin: 20px;
+}
+
+.clicker-button-container {
+    position:relative;
+    width: 50%;
+    display: flex;
+    justify-content: space-between;
+    margin: 40px;
+}
+
+.auto-clicker-button-container {
+    position:relative;
+    width: 50%;
+    display: flex;
+    justify-content: space-between;
+    margin: 40px;
+}
+#axe-button {
+    position: absolute;
+    font-family: 'Poppins', sans-serif;
+    font-size: 15px;
+    cursor: pointer;
+    color: black;
+    transform: translate(75%, 50%);
+    background-color: transparent;
+    border-radius: 45%;
+    /*border: none;*/
+}
+
+#auto-axe-button {
+    position: absolute;
+    font-family: 'Poppins', sans-serif;
+    font-size: 15px;
+    cursor: pointer;
+    color: black;
+    transform: translate(50%, 50%);
+    background-color: transparent;
+    border-radius: 45%;
+    /*border: none;*/
+}
+
+#wood-button {
+    cursor: pointer;
+}
+
+#wood-number, #wood-button {
+    text-align: center;
+    margin: 0 10px auto 10px;
+}
+
+.store {
+    background-color: #DC5F00;
+    text-align: center;
+    overflow: scroll;
+    height: 100%;
+    padding: 20px;
+    padding-bottom: 20px;
+}
+
+
+.post-contents {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    height: 100%;
+    width: 100%;
+}
+
+.post-image-container img {
+    max-height: 140px;
+    max-width: 200px;
+}
+  
+  .post-info-container {
+    width: 100%;
+    margin-top: 20px;
+}
+
+.post-title, .post-price {
+    color: white;
+}
+
+.post {
+    display: inline-block;
+    vertical-align: top;
+    height: 200px;
+    width: 200px;
+    margin: 10px;
+    padding: 10px;
+    border-radius: 8px;
+    background-color: black; 
+    box-shadow: 4px 4px 12px rgba(0, 0, 0, 0.25);
+    cursor: pointer;
+}
