+++ views/404.handlebars
@@ -1,0 +1,2 @@
+<h2>404</h2>
+<h3>Woops!  We couldn't find that page.</h3>
