+++ views/layouts/main.handlebars
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Benny's Clicker</title>
+
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Poppins&display=swap" rel="stylesheet">
+
+    <link rel="stylesheet" href="/styles.css">
+    <script src="/index.js" charset="utf-8" defer></script>
+</head>
+<body>
+    <header>
+        <h1 class="site-title"><a href="#">Benny's Clicker</a></h1>
+
+        <!-- <div class="right-aligned">
+            <h1 class="right-text"><a href="#">Save</a></h1>
+            <h1 class="right-text"><a href="#">Leaderboard</a></h1>
+        </div> -->
+    </header>
+
+    <main class="content">
+        {{{body}}}
+    </main>
+</body>
+</html>
