+++ views/partials/post.handlebars
@@ -1,0 +1,10 @@
+<div class="post" data-price="{{price}}">
+    <div class="post-contents">
+        <div class="post-image-container">
+            <img src="{{source}}" alt="{{itemDescription}}">
+        </div>
+        <div class="post-info-container">
+            <a href="#" class="post-title">{{itemDescription}}</a> <span class="post-price">${{price}}</span>
+        </div>
+    </div>
+</div>
